Split MavenDependencyResolver into Loader and Resolver
