+++ core/src/main/java/com/nikodoko/javaimports/resolver/maven/MavenDependencyResolver.java
@@ -1,74 +1,27 @@
 package com.nikodoko.javaimports.resolver.maven;
 
-import com.nikodoko.javaimports.parser.Import;
-import com.nikodoko.javaimports.resolver.ImportWithDistance;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.jar.JarEntry;
-import java.util.jar.JarInputStream;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-class MavenDependencyResolver {
-  private static final String SUBCLASS_SEPARATOR = "$";
-  // Supports "exotic" versioning, like guava's "26.0-jre"
-  private static final Pattern VERSION_REGEX =
-      Pattern.compile("(?:\\S+)?(?<version>\\d+(?:\\.\\d+)+)(?:\\S+)?");
-  final Path fileBeingResolved;
-  final Path repository;
+public class MavenDependencyResolver {
+  private final Path repository;
 
-  MavenDependencyResolver(Path fileBeingResolved, Path repository) {
-    this.fileBeingResolved = fileBeingResolved;
+  private MavenDependencyResolver(Path repository) {
     this.repository = repository;
   }
 
-  List<ImportWithDistance> load(Path dependency) throws IOException {
-    return scanJar(dependency);
+  static MavenDependencyResolver withRepository(Path repository) {
+    return new MavenDependencyResolver(repository);
   }
 
-  private List<ImportWithDistance> scanJar(Path jar) throws IOException {
-    List<ImportWithDistance> imports = new ArrayList<>();
-    try (JarInputStream in = new JarInputStream(new FileInputStream(jar.toString()))) {
-      JarEntry entry;
-      while ((entry = in.getNextJarEntry()) != null) {
-        // XXX: this will get all classes, including private and protected ones
-        // TODO: properly handle module-class.class
-        if (entry.getName().endsWith(".class")) {
-          Import i = parseImport(Paths.get(entry.getName()));
-          imports.add(ImportWithDistance.of(i, jar, fileBeingResolved));
-        }
-      }
-    }
-
-    return imports;
-  }
-
-  private Import parseImport(Path jarEntry) {
-    String pkg = jarEntry.getParent().toString().replace("/", ".");
-    String filename = jarEntry.getFileName().toString();
-    String name = filename.substring(0, filename.lastIndexOf("."));
-    if (!name.contains(SUBCLASS_SEPARATOR)) {
-      return new Import(name, pkg, false);
-    }
-
-    // Make the subclass addressable by its name
-    String extraPkg =
-        name.substring(0, name.lastIndexOf(SUBCLASS_SEPARATOR)).replace(SUBCLASS_SEPARATOR, ".");
-    String subclassName = name.substring(name.lastIndexOf(SUBCLASS_SEPARATOR) + 1, name.length());
-    return new Import(subclassName, String.join(".", pkg, extraPkg), false);
-  }
-
-  public Path resolve(MavenDependency dependency) throws IOException {
-    Path dependencyRepository =
-        Paths.get(
-            repository.toString(), dependency.groupId.replace(".", "/"), dependency.artifactId);
+  Path resolve(MavenDependency dependency) throws IOException {
+    Path dependencyRepository = directoryFor(dependency);
     String version = dependency.version;
     if (!dependency.hasPlainVersion()) {
       version = getLatestAvailableVersion(dependencyRepository).name;
@@ -76,6 +29,11 @@ class MavenDependencyResolver {
 
     return Paths.get(
         dependencyRepository.toString(), version, jarName(dependency.artifactId, version));
+  }
+
+  private Path directoryFor(MavenDependency dependency) {
+    return Paths.get(
+        repository.toString(), dependency.groupId.replace(".", "/"), dependency.artifactId);
   }
 
   private String jarName(String artifactId, String version) {
